feat: Improve media metadata parsing
`mime-types` was replaced with `mediainfo.js` which allows
us to get media type based on actual media contents
instead of relying on the filename. It also allows us
to get media dimensions and duration, which can be passed
to Telegram API request to fix annoying bugs when clients
display attachments invalidly.

This change required direct media from urls to be fetched
by us before passed to `mtcute`, so that was implemented
too.